pyproject.toml
[tool.poetry]
name = "todo-app"
version = "0.1.0"
description = "A simple ToDo app for your DevOps apprenticeship"
authors = []

[tool.poetry.dependencies]
python = "^3.9"
flask = "^1.1.2"
python-dotenv = "^0.14.0"
requests = "^2.25.1"
pytest = "^6.2.2"
selenium = "^3.141.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
